+++ src/config/styles.ts
@@ -63,8 +63,8 @@ const postcssLoaderOptions: PostcssLoaderOptions = {
 // CSS Sprites
 const imageBasePath = '../'; // Relative to css file
 const sprites: string[] = []; // Image folders
-const spriteParams: object = {};
 const spriteRetina = false;
+const spriteParams: object = {};
 
 export default {
   extname,
@@ -78,6 +78,6 @@ export default {
   postcssLoaderOptions,
   imageBasePath,
   sprites,
-  spriteParams,
-  spriteRetina
+  spriteRetina,
+  spriteParams
 };